requirements.thawed.txt
# THIS IS AN AUTO-GENERATED FILE. DO NOT EDIT!
flask
google-cloud-datastore
bcrypt
pymakehelper
pydmt
pylint
